refactor: プレゼンテーション層をリファクタリング（Phase 5 完了）
- presentation/hooks/: ゲームエンジン・入力・オーディオ・モバイルフックを作成
  - use-game-engine.ts: 25個の useState をフックに集約
  - use-input.ts: キーボード・タッチ入力の統合
  - use-audio.ts: AudioPort 経由のオーディオ制御
  - use-mobile.ts: モバイル判定
- presentation/screens/: 画面コンポーネントを分割
  - TitleScreen.tsx, PlayScreen.tsx, ResultScreen.tsx
- NonBrakeDescentGame.tsx を 632行 → 98行に縮小
  - 画面切替のみの薄いコンテナに変更
- 全 193 テストパス
