+++ src/features/non-brake-descent/presentation/hooks/index.ts
@@ -1,0 +1,10 @@
+/**
+ * プレゼンテーション層フックの re-export
+ */
+export { useGameEngine } from './use-game-engine';
+export type { UseGameEngineResult } from './use-game-engine';
+export { useIsMobile } from './use-mobile';
+export { useInput } from './use-input';
+export type { UseInputResult } from './use-input';
+export { useAudio } from './use-audio';
+export type { UseAudioResult } from './use-audio';
